Cache default values
Previously, if caching was turned on but a key was not yet in the persisted database,
the engine returned the default value without caching it.  Therefore, the next time
the key was requested, the engine attempted once more to find the value with a database call.

This commit makes the engine cache the default value so that the engine does not continue
to query the database every time the value is requested.